Make SuperMDS a Publishable Library
