+++ nb-configuration.xml
@@ -8,11 +8,11 @@ Without this configuration present, some functionality in the IDE may be limited
 -->
     <properties xmlns="http://www.netbeans.org/ns/maven-properties-data/1">
         <!--
-Properties that influence various parts of the IDE, especially code formatting and the like. 
+Properties that influence various parts of the IDE, especially code formatting and the like.
 You can copy and paste the single properties, into the pom.xml file and the IDE will pick them up.
 That way multiple projects can share the same settings (useful for formatting rules for example).
 Any value defined here will override the pom.xml file value but is only applicable to the current project.
 -->
-        <netbeans.hint.jdkPlatform>JDK_23</netbeans.hint.jdkPlatform>
+        <netbeans.hint.jdkPlatform>JDK_21</netbeans.hint.jdkPlatform>
     </properties>
 </project-shared-configuration>
