+++ src/main/java/module-info.java
@@ -1,0 +1,6 @@
+module com.github.trinity.supermds {
+    requires commons.math3;
+    requires org.slf4j;
+
+    exports  com.github.trinity.supermds;
+}
